docker/pc2_infra_core/requirements.txt
# PC2 Infrastructure Core Requirements - Base packages
redis==4.6.0
nats-py==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
psutil==5.9.6
aiohttp==3.9.1
asyncio-mqtt==0.13.0
pydantic==2.5.2
python-dotenv==1.0.0
PyJWT==2.8.0
websockets==12.0
requests==2.31.0

# ZMQ and YAML support
pyzmq==25.1.2
PyYAML==6.0.1

# Additional common dependencies (sqlite3 is built into Python)
sqlalchemy==2.0.25
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Optional GPU/ML libs used by some agents (ResourceManager GPU monitoring)
torch==2.1.0

# HTTP and async support
httpx==0.25.2
aiofiles==23.2.1
starlette==0.27.0
